refactor(tii): extract schema interpretation into its own module
Pure readability rework of the complex param-type interpreter; no behavior
change (identical kind map, all existing tests pass).

- Move ParamType / VariantCase / ParamMap / params_from_schema and the
  from_json_schema interpreter out of the 886-line mod.rs into a new
  self-contained schema.rs; re-export so tii::ParamType etc. stay stable.
- Split the ~60-line from_json_schema into a thin dispatcher plus named
  helpers (ref_type, variant_type, array_type, object_type).
- Drop the now-dead Error::InvalidParamType / InvalidParamsSchema variants
  (from_json_schema no longer returns Result).
- Add an invoke-level integration test over the shared complex.tii fixture
  that exercises every kind incl. component-resolved record/variant, locking
  in the components threading through Protocol::invoke.
